Move Tril AST implementation to C++ file
A C header now contains forward declarations to allow C to interact with
the AST. The implementation is in C++.
