+++ fvtest/tril/test/ParserTest.cpp
@@ -18,7 +18,7 @@
 
 #include "gtest/gtest.h"
 
-#include "ast.h"
+#include "ast.hpp"
 
 #define ASSERT_NULL(pointer) ASSERT_EQ(nullptr, (pointer))
 #define ASSERT_NOTNULL(pointer) ASSERT_TRUE(nullptr != (pointer))
