+++ fvtest/tril/tril/ast.hpp
@@ -1,0 +1,47 @@
+/*******************************************************************************
+ *
+ * (c) Copyright IBM Corp. 2017, 2017
+ *
+ *  This program and the accompanying materials are made available
+ *  under the terms of the Eclipse Public License v1.0 and
+ *  Apache License v2.0 which accompanies this distribution.
+ *
+ *      The Eclipse Public License is available at
+ *      http://www.eclipse.org/legal/epl-v10.html
+ *
+ *      The Apache License v2.0 is available at
+ *      http://www.opensource.org/licenses/apache2.0.php
+ *
+ * Contributors:
+ *    Multiple authors (IBM Corp.) - initial implementation and documentation
+ ******************************************************************************/
+
+#ifndef AST_HPP
+#define AST_HPP
+
+#include "ast.h"
+
+struct ASTValue {
+    ASTValueType type;
+    union {
+        uint64_t int64;
+        double f64;
+        const char* str;
+    } value;
+    ASTValue* next;
+};
+
+struct ASTNodeArg {
+    const char* name;
+    ASTValue* value;
+    ASTNodeArg* next;
+};
+
+struct ASTNode {
+    const char* name;
+    ASTNodeArg* args;
+    ASTNode* children;
+    ASTNode* next;
+};
+
+#endif // AST_HPP
